Update user_data.json
Example of what the user data file might look like. This file is required for the sign-in module to work, as the sign-in module will check the entered username and password against this file, and make a determination on what feedback should be given to the user (e.g: "sign-in successful" or "incorrect password".